requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-restframework==0.0.1
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0


certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
django-allauth==0.44.0
django-colorfield==0.4.1

django-rest-auth==0.9.5

djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
oauthlib==3.1.0

pyclean==2.0.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0

requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0

urllib3==1.26.4

dj-database-url
python-decouple
python-dotenv